feat: refactor GroupModel import paths and update method definitions to arrow functions for consistency
