+++ libs/ui/recursive-group-list/src/components/recursive-group-list/recursive-group-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ContextMenuComponent } from '../context-menu/context-menu.component';
-import { GroupModel } from '../../models/group-models';
+import { GroupModel } from '../../models/group-model';
 
 @Component({
   selector: 'clx-recursive-group-list',
@@ -31,22 +31,16 @@ export class RecursiveGroupListComponent {
     this.contextMenuVisible = true;
     this.contextMenuPosition = { x: event.clientX, y: event.clientY };
 
-    this.showSelectedGroup();
-
     console.log(`Updated selectedGroup for menu: ${this.selectedGroup.id} - ${this.selectedGroup.name}`);
   }
 
-  showSelectedGroup() {
-    console.log(`Showing Selected group: ${this.selectedGroup?.id} - ${this.selectedGroup?.name}\n\r ${this.selectedGroup?.subGroups}`);
-  }
-
-  renameGroup(newName: string, group: GroupModel) {
+  renameGroup = (newName: string, group: GroupModel) => {
     console.log(`Renaming group: ${group.id} - ${group.name} to: ${newName.trim()}`);
     this.renameGroupById(group.id, newName.trim());
-    console.log(`Group renamed to: ${ newName.trim() }`);
+    console.log(`Group renamed to: ${newName.trim()}`);
   }
 
-  deleteGroup(groupToDelete: GroupModel) {
+  deleteGroup = (groupToDelete: GroupModel) => {
     console.log(`Deleting group: ${groupToDelete.id} - ${groupToDelete.name}`);
 
     const removeGroupById = (groups: GroupModel[]): GroupModel[] => {
@@ -63,28 +57,18 @@ export class RecursiveGroupListComponent {
     console.log(`Group deleted successfully: ${groupToDelete.id}`);
   }
 
-
-  // deleteGroup(group: GroupModel) {
-  //   console.log(`Deleting selected group:\n\r ${group.id}`);
-  //   const groupIndex = this.groups.findIndex(g => g.id === group.id);
-  //   if (groupIndex === -1) return;
-
-  //   this.groups.splice(groupIndex, 1);
-  //   console.log(`Group deleted: ${group.id}`);
-  // }
-
-  toggleGroup(group: GroupModel) : void {
-    console.log(`Group toggled in model: ${ group.id }`);
+  toggleGroup = (group: GroupModel): void => {
+    console.log(`Group toggled in model: ${group.id}`);
     group.expanded = !group.expanded;
 
     // Recursively collapse all child groups if this group is collapsed
     // If collapsing, set all child groups' 'expanded' to false
     if (!group.expanded && (group.subGroups?.length ?? 0) > 0) {
-        this.setAllChildGroupsCollapsed(group);
+      this.setAllChildGroupsCollapsed(group);
     }
   }
 
-  renameGroupById(groupId: string, newName: string): void {
+  renameGroupById = (groupId: string, newName: string): void => {
     console.log(`Renaming group with ID: ${groupId} to '${newName}'`);
 
     let groupFound = false;
@@ -109,24 +93,24 @@ export class RecursiveGroupListComponent {
     }
   }
 
-  isGroup(group: GroupModel): boolean {
+  isGroup = (group: GroupModel): boolean => {
     return !!group.subGroups && (group.subGroups?.length ?? 0) > 0;
   }
 
-  isExpanded(group: GroupModel): boolean {
-      return group.expanded;
+  isExpanded = (group: GroupModel): boolean => {
+    return group.expanded;
   }
 
   // Helper method to set all child groups to collapsed
-  private setAllChildGroupsCollapsed(group: GroupModel): void {
-    if(!!group.subGroups)
+  private setAllChildGroupsCollapsed = (group: GroupModel): void => {
+    if (!!group.subGroups)
       group.subGroups.forEach(subGroup => {
-          subGroup.expanded = false;
+        subGroup.expanded = false;
 
-          // If the child group has its own children, collapse them as well
-          if (!!subGroup.subGroups && subGroup.subGroups.length > 0) {
+        // If the child group has its own children, collapse them as well
+        if (!!subGroup.subGroups && subGroup.subGroups.length > 0) {
           this.setAllChildGroupsCollapsed(subGroup);
-          }
+        }
       });
   }
 }
