+++ libs/ui/recursive-group-list/src/models/group-model.ts
