+++ src/models/index.ts
@@ -1,1 +1,1 @@
-export * from "../../libs/ui/recursive-group-list/src/models/group-models";
+export * from "../../libs/ui/recursive-group-list/src/models/group-model";
